fix: resolve CI container blocks and finalize merch design pipeline
- Migrate CI Build & E2E, snapshot updates, and WCS ETL workflows to host runner to avoid MCR registry blocks.
- Refactor merch design generation to use direct Cairo paths for Heart and Checkmark, ensuring 6-color brand compliance and exact visual fidelity.
- Align rainbow stripes precisely to shape boundaries.
- Clean up intermediate SVG assets and binary screenshots.
- Verified all assets via automated QA script.
